Refactor metadata management by introducing MetadataProducer class for streamlined extraction of Composition sections and their constraints; remove unused constants and methods from Generator class.
Ref: https://github.com/hl7au/au-ps-inferno/issues/97
